[BUG] Check wpcom_vip function exists before using it
Check that `wpcom_vip_file_put_contents` function exists before trying
to use it.

I can't find any documentation on this function, but presumably it's for
working with VIP's filesystem. Although the beginning of this
`export_log` method does seem to suggest that this functionality won't
work on VIP environments anyway...

Currently there is a fatal error when I try to process one of the GDPR
requests on my local environment.

This `wpcom_vip` function was added in a commit for "PHPCS fixes" -

https://github.com/boxuk/GDPR/pull/40/commits/eabd9f14341c11be2657db8f6b49c45f543027c9#diff-ea4ff831e9f6ef28b1793cd99973b9947370d9c3603059927d5c99487db97987L174